Rename docker compose settings file
